+++ lilleprinsen-click-collect/lilleprinsen-click-collect.php
@@ -35,6 +35,7 @@ require_once LP_CC_PLUGIN_PATH . 'includes/class-compatibility.php';
 require_once LP_CC_PLUGIN_PATH . 'includes/class-activator.php';
 require_once LP_CC_PLUGIN_PATH . 'includes/class-deactivator.php';
 require_once LP_CC_PLUGIN_PATH . 'includes/class-plugin.php';
+require_once LP_CC_PLUGIN_PATH . 'includes/class-staff-profiles.php';
 require_once LP_CC_PLUGIN_PATH . 'includes/class-settings.php';
 require_once LP_CC_PLUGIN_PATH . 'includes/class-pickup-number.php';
 require_once LP_CC_PLUGIN_PATH . 'includes/class-qr-code.php';